Refactor: Auth 컨트롤러 response 객체 적용 #4 Rename: validator 파일 컨벤션에 맞게 파일명 수정 #4
