+++ src/auth/auth.router.js
@@ -2,8 +2,8 @@ import express from "express";
 import { AuthController } from "./auth.controller.js";
 import { AuthService } from "./auth.service.js";
 import { prisma } from "../configs/prisma-client.js";
-import { signUpValidator } from "../middlewares/validators/signUp-validator.js";
-import { signInValidator } from "../middlewares/validators/signIn-validator.js";
+import { signUpValidator } from "../middlewares/validators/sign-up-validator.middleware.js";
+import { signInValidator } from "../middlewares/validators/sign-in-validator.middleware.js";
 
 const authRouter = express.Router();
 const authService = new AuthService(prisma);
